Delete Property functionality with Authorization

+++ controllers/propertyController.js
@@ -13,6 +13,13 @@ class PropertyController {
       locationDetails,
       city,
     } = req.body;
+    const propertyImagesArray = req.files["propertyImage"];
+    // console.log(propertyImagesArray)
+    var imagesArray = [];
+    propertyImagesArray.forEach((element) => {
+      imagesArray.push(element.filename);
+    });
+    // console.log(imagesArray)
     if (
       (title && description && priceCoin,
       priceDes,
@@ -35,6 +42,7 @@ class PropertyController {
             details: locationDetails,
           },
           city: city,
+          photos: imagesArray,
         });
 
         await doc.save();
@@ -44,6 +52,7 @@ class PropertyController {
           message: "Property Listed for sale",
         });
       } catch (err) {
+        // console.log(err)
         res.status(201).send({
           status: "failed",
           message: "DB Error",
@@ -60,31 +69,53 @@ class PropertyController {
   static getPropertyDetails = async (req, res) => {
     const { id } = req.params;
     try {
-      const property = await PropertyModel.findById(id).populate({path:'ownerId',select:'-password'})
+      const property = await PropertyModel.findById(id).populate({
+        path: "ownerId",
+        select: "-password",
+      });
       if (property) {
         res.send(property);
       } else {
         res.send("Not Found");
       }
     } catch (err) {
-      res.send("DB Error"+ err);
+      res.send("DB Error" + err);
     }
   };
 
-  static deleteProperty = async (req, res)=>{
+  static deleteProperty = async (req, res) => {
     const { id } = req.params;
+    const userEmail = req.user.email;
+    console.log(userEmail);
     try {
-      const property = await PropertyModel.findByIdAndDelete(id)
-      console.log(property)
-      if (property) {
-        res.send(property);
-      } else {
-        res.send("Not Found");
+      const checkPropertyAuth = await PropertyModel.findById(id).populate(
+        "ownerId"
+      );
+      // id porperty Found then 
+      if (checkPropertyAuth) {
+        if (checkPropertyAuth.ownerId.email === userEmail) {
+          const property = await PropertyModel.findByIdAndDelete(id);
+          // console.log(property);
+          if (property) {
+            res.send({status: "success",
+            message: "Property Listed for sale", "property": property});
+          } else {
+            res.send({status: "failed",
+            message: "Property already Deleted!",});
+          }
+        } else {
+          res.send({status: "failed",
+          message: "You are not allowed",});
+        }
+      }else{
+        res.send({status: "failed",
+        message: "Property not found",});
       }
+      // console.log(checkPropertyAuth.ownerId.email)
     } catch (err) {
       res.send("DB Error");
     }
-  }
+  };
 }
 
 export default PropertyController;
